pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ceus-lrm-fusion"
version = "0.1.0"
description = "Public reference implementation for CEUS-GRU, Clinical-LR, and LRM-Fusion."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "joblib>=1.3",
  "matplotlib>=3.8",
  "numpy>=1.24",
  "openpyxl>=3.1",
  "pandas>=2.0",
  "PyYAML>=6.0",
  "scikit-learn>=1.3",
  "torch>=2.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
